requirements.txt
# LangChain & LangGraph - Core AI framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0

# PDF Processing
pdfplumber>=0.10.0

# Data Validation
pydantic>=2.0.0

# Web Interface
gradio>=4.0.0

# Environment Variables
python-dotenv>=1.0.0
